# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de controle financeiro pessoal com autenticação JWT",
        "title": "Imperium API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tron/projects": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all TRON projects for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tron-projects"
                ],
                "summary": "List all TRON projects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TronProjectResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new TRON project for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tron-projects"
                ],
                "summary": "Create a new TRON project",
                "parameters": [
                    {
                        "description": "Project data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTronProjectRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.TronProject"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tron/projects/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a specific TRON project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tron-projects"
                ],
                "summary": "Get a TRON project by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TronProjectResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a specific TRON project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tron-projects"
                ],
                "summary": "Update a TRON project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTronProjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TronProject"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a specific TRON project and all associated data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tron-projects"
                ],
                "summary": "Delete a TRON project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateTronProjectRequest": {
            "type": "object",
            "properties": {
                "daily_budget": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "frequency": {
                    "$ref": "#/definitions/models.TronFrequency"
                },
                "name": {
                    "type": "string"
                },
                "references": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "models.TronFrequency": {
            "type": "string",
            "enum": [
                "normal",
                "high",
                "low"
            ],
            "x-enum-varnames": [
                "FrequencyNormal",
                "FrequencyHigh",
                "FrequencyLow"
            ]
        },
...
        "models.TronProject": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "daily_budget": {
                    "description": "Max daily spend in USD",
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "directives": {
                    "description": "IDs of TronDirective",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "frequency": {
                    "description": "How often to run cycles",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TronFrequency"
                        }
                    ]
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "description": "Whether agents should run",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "references": {
                    "description": "Market references (competitor URLs)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "repos": {
                    "description": "IDs of TronRepo",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.TronProjectResponse": {
            "type": "object",
            "properties": {
                "active_directive": {
                    "type": "string"
                },
                "commits_today": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "daily_budget": {
                    "description": "Max daily spend in USD",
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "directives": {
                    "description": "IDs of TronDirective",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "frequency": {
                    "description": "How often to run cycles",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TronFrequency"
                        }
                    ]
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "description": "Whether agents should run",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "references": {
                    "description": "Market references (competitor URLs)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "repos": {
                    "description": "IDs of TronRepo",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "repos_count": {
                    "type": "integer"
                },
                "tasks_completed": {
                    "type": "integer"
                },
                "tasks_in_backlog": {
                    "type": "integer"
                },
                "today_cost_usd": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateTronProjectRequest": {
            "type": "object",
            "properties": {
                "daily_budget": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "frequency": {
                    "$ref": "#/definitions/models.TronFrequency"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "references": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite: Bearer {seu_token_aqui}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}